refactor: create utility for nested templates
